refactor: move chat api helpers into chat feature
