+++ client/src/features/chat/pages/chat-workout-draft-page.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen, waitFor, within } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { beforeEach, describe, expect, it, vi } from "vitest";
-import type { AIWorkoutDraft } from "@/lib/api/ai-chat";
+import type { AIWorkoutDraft } from "@/features/chat/api/ai-chat";
 import {
   ChatRouteComponent,
   conversationDetail,
