+++ client/src/features/chat/utils/ai-chat-observability.test.ts
