+++ client/src/features/chat/utils/ai-workout-draft.ts
@@ -2,7 +2,7 @@ import type {
   WorkoutCreateWorkoutRequest,
   WorkoutExerciseInput,
 } from "@/client";
-import type { AIWorkoutDraft } from "@/lib/api/ai-chat";
+import type { AIWorkoutDraft } from "@/features/chat/api/ai-chat";
 import {
   type WorkoutDraftStorage,
   workoutDraftStorage,
