+++ client/src/features/chat/utils/chat-session-submit.ts
@@ -2,12 +2,12 @@ import {
   createAIChatConversation,
   streamAIChatMessage,
   type AIChatMessage,
-} from "@/lib/api/ai-chat";
+} from "@/features/chat/api/ai-chat";
 import {
   classifyRecoveryOutcome,
   classifyStreamInterruption,
   terminalStreamStage,
-} from "@/lib/ai-chat-observability";
+} from "@/features/chat/utils/ai-chat-observability";
 import { getErrorMessage, showErrorToast } from "@/lib/errors";
 import {
   clearResumeCursor,
